+++ pkg/providers/cloudcertificates/testdata/TestDataCertificate/certificate_id_is_missing.tf
@@ -1,0 +1,6 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_cloudcertificates_certificate" "testcert" {
+}