+++ pkg/providers/cloudcertificates/testdata/TestDataCertificate/certificate_with_bindings.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_cloudcertificates_certificate" "testcert" {
+  certificate_id            = "12345"
+  include_hostname_bindings = true
+}